Rename deleteEmptyValuesFromObject to deleteEmptyPropertiesFromObject
