+++ src/domain/events/utils.ts
@@ -6,14 +6,14 @@ import {
   EventsFieldsFragment,
   EventsQuery,
 } from '../../generated/graphql';
-import deleteEmptyValuesFromObject from '../../utils/deleteEmptyValuesFromObject';
+import deleteEmptyPropertiesFromObject from '../../utils/deleteEmptyPropertiesFromObject';
 import { queryParameterToArray } from '../../utils/queryParameterToArray';
 import { EventSearchFormValues } from './eventSearchForm/EventSearchForm';
 
 export const getSearchQueryObject = (
   values: EventSearchFormValues
 ): Record<string, string | string[]> => {
-  return deleteEmptyValuesFromObject({
+  return deleteEmptyPropertiesFromObject({
     ...values,
     date: values.date?.toISOString(),
     endDate: values.endDate?.toISOString(),
