+++ src/utils/deleteEmptyPropertiesFromObject.ts
